@@ -13,32 +13,32 @@
     "author": "Philip Washington Sorst <philip@sorst.net> (https://sorst.net)",
     "private": true,
     "dependencies": {
-        "@angular/animations": "^7.2.11",
-        "@angular/cdk": "^7.3.6",
-        "@angular/common": "^7.2.11",
-        "@angular/compiler": "^7.2.11",
-        "@angular/core": "^7.2.11",
-        "@angular/forms": "^7.2.11",
+        "@angular/animations": "^8.2.14",
+        "@angular/cdk": "^8.2.3",
+        "@angular/common": "^8.2.14",
+        "@angular/compiler": "^8.2.14",
+        "@angular/core": "^8.2.14",
+        "@angular/forms": "^8.2.14",
         "@angular/http": "^7.2.11",
-        "@angular/material": "^7.3.6",
-        "@angular/platform-browser": "^7.2.11",
-        "@angular/platform-browser-dynamic": "^7.2.11",
+        "@angular/material": "^8.2.3",
+        "@angular/platform-browser": "^8.2.14",
+        "@angular/platform-browser-dynamic": "^8.2.14",
         "@angular/pwa": "0.12",
-        "@angular/router": "^7.2.11",
-        "@angular/service-worker": "^7.2.11",
+        "@angular/router": "^8.2.14",
+        "@angular/service-worker": "^8.2.14",
         "core-js": "^2.5.3",
         "faker": "^4.1.0",
         "hammerjs": "^2.0.8",
         "material-design-icons": "^3.0.1",
-        "rxjs": "^6.4.0",
+        "rxjs": "^6.6.3",
         "tslib": "^1.9.0",
-        "zone.js": "^0.8.20"
+        "zone.js": "^0.9.1"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "~0.13.0",
-        "@angular/cli": "^7.3.7",
-        "@angular/compiler-cli": "^7.2.11",
-        "@angular/language-service": "^7.2.11",
-        "typescript": "~3.2.4"
+        "@angular-devkit/build-angular": "~0.803.29",
+        "@angular/cli": "^8.3.29",
+        "@angular/compiler-cli": "^8.2.14",
+        "@angular/language-service": "^8.2.14",
+        "typescript": "~3.5.3"
     }
 }
